Annual Summary Shapefile .py Cleanup
Code development and cleanup.
Added input arguments.
Included both median and mean output stats.
General style corrections.
Renamed files and removed developmet scripts (_old).
